requirements.txt
flask
pymysql
pandas
numpy
python-dotenv
cryptography
